Move typography components to flattened structure
