+++ packages/components/src/index.ts
@@ -9,6 +9,8 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
+export * from './Bar'
+export * from './BulletList'
 export * from './Checkbox'
 export * from './Content'
 export * from './Currency'
@@ -19,7 +21,10 @@ export * from './FormTabs'
 export * from './Frame'
 export * from './ImageUploader'
 export * from './InputField'
+export * from './Legend'
 export * from './LineChart'
+export * from './Link'
+export * from './List'
 export * from './PDFViewer'
 export * from './PageWrapper'
 export * from './Pagination'
@@ -32,6 +37,7 @@ export * from './SubSectionDivider'
 export * from './Text'
 export * from './TextArea'
 export * from './TextInput'
+export * from './VerticalBar'
 export * from './WarningMessage'
 export * from './Workqueue'
